Community widget gallery (#225)
This adds a new way of adding custom widgets and introduces the concept of
asset galleries (currently limited to widgets), with a first implementation of
a gallery retrieving the widgets posted on the community.openhab.org
Discourse instance.

The gallery replaces the GitHub import feature but retains all the previous
functionality. In fact if a community post references a GitHub repo instead
of directly attaching .widget.json files in the post, the gallery will provide
a button to navigate to the repo and import the widgets from there.

Widgets can also be added directly from the dashboard editor using the new
"Get more..." link in the "Add Widget" dropdown menu.

Widgets imported from GitHub or community.openhab.org now have a "Help" button
in their settings leading to resp. the GitHub repo readme or the forum topic.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>